requirements.txt
streamlit>=1.29.0
pandas>=2.0.0
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.15.0
openpyxl>=3.1.0

# Data fetching
requests>=2.31.0
python-dotenv>=1.0.0
yfinance>=0.2.36
pyarrow>=14.0.0

# ML Models
prophet>=1.1.5
xgboost>=2.0.0
scikit-learn>=1.3.0

# Snowflake (uncomment when ready to deploy)
# snowflake-connector-python>=3.6.0
